Read account data using specialized routes
Now that matrix-org/synapse#4303 is implemented,
we don't need to resort to doing `/sync` to
retrieve account data.

We can do it via the new GET endpoint of the API:
`/user/{user_id}/account_data/{account_dataType}`

This API endpoint landed in Synapse v0.34.1,
so versions prior to that will not be supported anymore.