Change HTTP Gateway timeout from 15 sec. to Matrix.TimeoutMilliseconds
Some requests (like `/sync`) take a long time on the Synapse side,
either because Synapse is slow to respond or because they're
long-polling requests with a timeout of 30 seconds.

The default timeout (15 sec.) we had for the HTTP Gateway was not enough
and requests were often aborted prematurely.

With this change, we're rausing the HTTP Gateway's timeout value to
the same (generally high-enough) value as `Matrix.TimeoutMilliseconds`.