Initial work on an event/hook system
This is very new and experimental and the details around it
may change soon.

The goal is to provide a rule-based firewall system like
[mxgwd](https://github.com/kamax-matrix/mxgwd), except more flexible.

Currently only the first hook that matches runs.
In the future, it might make sense to extend it.

Besides being able to filter out any request via a static policy,
we also support consulting a remote REST service.

We do provide an example of such a service (written in PHP) in
`etc/services/hook-rest-service/index.php` (reachable via
`http://hook-rest-service:8080` in a container
and `http://localhost:41466` when running locally).

Such a REST service can also be used for logging, etc.

Besides rejecting or delegating request checking to a service,
we also support response modification (statically or via a REST
service). A hook can request that additional fields (or headers)
are injected into the response.