Make room event-sending policy-checked (always allowed)
We intercept room event-sending requests now, because:

- it might be useful as a future feature (controlling it via the policy)

- making it a "policy-checked request" (and not a catch-all one)
allows people to use the
`EventTypeBeforeAuthenticatedPolicyCheckedRequest`
("beforeAuthenticatedPolicyCheckedRequest") event hook with it. It would
be great if we could such "authenticated-user" hooks for catch-all
requests, but it's not implemented yet. Making this API policy-checked
allows this hook to be used there, until we think of a better way