Auto-expire (access token -> user id) resolver data
The fact that some token is valid now, does not necessarily mean
that it's valid forever.

With a large enough LRU cache size, it's possible that we'd
never expire a (token to user id) resolver result and we'd
always consider that a given token is owned by a given user.

This could be misleading in logs, but it could also mislead
other systems (REST service hooks).

Expiration time can be controlled now. People who wish
to ensure (with higher certainty) that a given access token
is really owned (and valid) for a given user, can now do so,
by lowering the expiration time value.

Of course, this comes at the expense of performance,
as we need to hit `/.../account/whoami` each time we encounter
an expiration.

Even without tweaking the value, we do provide better defaults now.
A 5 minute expiration is likely a good balance.