Remove communities/groups support
Synapse v1.61.0 completely removes support for communities/groups.
Trying to use previous versions of matrix-corporal (which try to hit the community APIs)
fails.

This patch completely removes support for communities from
matrix-corporal as well, making it possible to use it with newer Synapse
versions (v1.61.0+), as well as older Synapse versions.

It may be beneficial to add support for the new Spaces functionality,
but that is yet unplanned. It may happen later. Spaces are just Rooms
though, so it may be possible to make use of spaces in some ways even
now.