Pin woodpeckerci/plugin-docker-buildx version and switch to docker-in-docker
For some reason, building previously worked even without docker-in-docker.
Supposedly, it was using Docker on the host, the socket of which is
exposed to the Woodpecker CI agent container by default.

A few things have changed since then:

- Woodpecker CI (server & agent) got upgraded (v1 -> v2)
- the host OS where Woodpecker CI agent runs (in a container) changed

It seems like plugin-docker-buildx currently cannot use Docker on the
host (without additional work) and tries to start its own Docker daemon and fails.

----

I've found that we have 2 choices to make building work again:

1) (Our choice with this patch) Use a separate docker-in-docker service.

This is clean and seems to work well. The benefit of using a throwaway
docker-in-docker is that after building, all temporary data will be removed.

We need to mark the repository as Trusted though.

2) Use Docker from the host

We can use Docker from the host by using a `volumes` settings key in the
pipeline, to mount `/var/run/docker.sock` to `/var/run/docker.sock` in
the `woodpeckerci/plugin-docker-buildx` container.

Mounting volumes also requires that the repository is marked as Trusted
though, so there's no benefit to switching to this.
The only benefits are:

- the cleaner pipeline history (no dind dependency service to start)
- a small speedup due to not having to start one more Docker daemon
  which also has nothing cached in it